move tests to single class
